Flight List Results
We retrieve the flights and publish on a new page with all the dates.

Futher refactoring can improve models to recieve Date() instead of Strings. Also maybe renaming some variables to the same ideas (departure, origin, initial should be renamed to the same name...)
Added a correction to station search so it search based on query instead of lowercase

Maybe visual improvements could be better

This is a working version